+++ VoxelEngine/VoxelEngine/VoxelEngine/CoreEngine/ParticleSystem.cpp
@@ -70,7 +70,7 @@ void ParticleSystem::Update(Scene& scene, float dt)
     {
 #pragma region uniforms
       emitter_shader->setInt("u_particlesToSpawn", particlesToSpawn);
-      emitter_shader->setFloat("u_time", timer.elapsed() + 1.61803);
+      emitter_shader->setFloat("u_time", static_cast<float>(timer.elapsed()) + 1.61803f);
       emitter_shader->setMat4("u_model", transform.GetModel());
       emitter_shader->setFloat("u_emitter.minLife", emitter.minLife);
       emitter_shader->setFloat("u_emitter.maxLife", emitter.maxLife);
